test(compaction): move #1667 budget-honesty TDD tests to integration file (campsite)
